+++ src/js/ui/NewDataExploration/LayerNavigationControl.js
@@ -1,15 +1,13 @@
-import React from 'react';
+import React, {useState} from 'react';
 import { makeStyles } from '@material-ui/core/styles';
-import DESearchBar from './DESearchBar';
 import {Button, ButtonGroup, Grid} from "@material-ui/core";
-import ClearIcon from '@material-ui/icons/Clear';
 import SaveIcon from '@material-ui/icons/Save';
 import FolderOpenIcon from '@material-ui/icons/FolderOpen';
-import AddIcon from "@material-ui/icons/Add";
+import SearchBarSection from "./SearchBarSection";
+import UpdateIcon from '@material-ui/icons/Update';
 
 //FIXME Replace hard coded number with this once datasets load immediately
-import {numberOfDatasets} from "../TabSystem";
-import CustomSearchbar from "./CustomSearchbar";
+import {layerTitles, numberOfDatasets} from "../TabSystem";
 
 const useStyles = makeStyles((theme) => ({
     root: {
@@ -24,44 +22,43 @@ const useStyles = makeStyles((theme) => ({
     },
 }));
 
+function buildWorkspace(selectedDatasets) {
+    let newBooleanWorkspace = [];
+    for(let i = 0; i < layerTitles.length; i++) {
+        if(selectedDatasets.includes(layerTitles[i])) {
+            newBooleanWorkspace.push(true);
+        }
+        else {
+            newBooleanWorkspace.push(false);
+        }
+    }
+    return newBooleanWorkspace;
+}
+
 export default function LayerNavigationControl(props) {
     const classes = useStyles();
+    const [selectedDatasets, setSelectedDatasets] = useState([]);
 
-    if(props.isWorkspace) {
-        return (
-            <div className={classes.root}>
-                <DESearchBar openLayers={props.openLayers} setOpenLayers={props.setOpenLayers}
-                             booleanWorkspace={props.booleanWorkspace} setBooleanWorkspace={props.setBooleanWorkspace}
-                             setSearchedDatasets={props.setSearchedDatasets}/>
-                <Grid container direction="row" justify="center" alignItems="center">
-                    <ButtonGroup>
-                        <Button variant="outlined" onClick={() => {
-                            const temp = new Array(19).fill(false);
-                            props.setBooleanWorkspace(temp);
-                        }} startIcon={<ClearIcon/>}>Clear Workspace</Button>
-                        <Button variant="outlined" startIcon={<SaveIcon />}>Save Workspace</Button>
-                        <Button variant="outlined" startIcon={<FolderOpenIcon />}>Load Workspace</Button>
-                    </ButtonGroup>
-                </Grid>
-            </div>
-        )
-    }
-
-    else {
-        return(
-            <div className={classes.root}>
-                <DESearchBar openLayers={props.openLayers} setOpenLayers={props.setOpenLayers}
-                             booleanWorkspace={props.booleanWorkspace} setBooleanWorkspace={props.setBooleanWorkspace}
-                             setSearchedDatasets={props.setSearchedDatasets}/>
-                {/*<CustomSearchbar setSearchedDatasets={props.setSearchedDatasets} />*/}
-                <Grid container direction="row" justify="center" alignItems="center">
-                    <Button variant="outlined" startIcon={<AddIcon />} onClick={() => {
-                        const temp = new Array(19).fill(true);
-                        props.setBooleanWorkspace(temp);
-                    }}>Add All Datasets To Workspace</Button>
-                </Grid>
-            </div>
-        )
-    }
-
+    return (
+        <div className={classes.root}>
+            <Grid container direction="row" justify="center" alignItems="center">
+                <ButtonGroup>
+                    <Button
+                        variant="outlined"
+                        onClick={() => props.setBooleanWorkspace(buildWorkspace(selectedDatasets))}
+                        startIcon={<UpdateIcon/>}
+                    >Update Workspace</Button>
+                    {/*<Button variant="outlined" onClick={() => {*/}
+                    {/*    const temp = new Array(19).fill(false);*/}
+                    {/*    props.setBooleanWorkspace(temp);*/}
+                    {/*}} startIcon={<ClearIcon/>}>Clear Workspace</Button>*/}
+                    <Button variant="outlined" startIcon={<SaveIcon />}>Save Workspace</Button>
+                    <Button variant="outlined" startIcon={<FolderOpenIcon />}>Load Workspace</Button>
+                </ButtonGroup>
+            </Grid>
+            <SearchBarSection selectedDatasets={selectedDatasets} setSelectedDatasets={setSelectedDatasets}
+                openLayers={props.openLayers} setOpenLayers={props.setOpenLayers}
+                         booleanWorkspace={props.booleanWorkspace} setBooleanWorkspace={props.setBooleanWorkspace}/>
+        </div>
+    )
 }
